+++ vitest.config.ts
@@ -1,14 +1,22 @@
-import { defineConfig } from 'vitest/config'
+import { defineConfig } from "vitest/config";
+import react from "@vitejs/plugin-react";
 
 export default defineConfig({
+    plugins: [
+        react({
+            babel: {
+                plugins: ["babel-plugin-react-compiler"],
+            },
+        }),
+    ],
 
-  test: {
-    globals: true,     // so can use `describe`/`it` without importing them
-    environment: 'node', 
-    include: ['tests/**/*.test.{js,ts}'], 
-  },
+    test: {
+        globals: true, // so can use `describe`/`it` without importing them
+        environment: "node",
+        include: ["tests/**/*.test.{js,ts}"],
+    },
 
-  define: {
-    __VERSION__: JSON.stringify('test-version'),
-  }
-})
+    define: {
+        __VERSION__: JSON.stringify("test-version"),
+    },
+});